+++ backend/src/main/java/kakaobootcamp/backend/domains/email/service/EmailTokenService.java
@@ -1,22 +1,32 @@
-package kakaobootcamp.backend.domains.email;
+package kakaobootcamp.backend.domains.email.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import kakaobootcamp.backend.common.exception.ApiException;
 import kakaobootcamp.backend.common.exception.ErrorCode;
+import kakaobootcamp.backend.domains.email.domain.EmailToken;
 import kakaobootcamp.backend.domains.email.repository.EmailTokenRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class EmailVerificationService {
+public class EmailTokenService {
 	private final EmailTokenRepository emailTokenRepository;
+
+	// 인증된 이메일 저장
+	@Transactional
+	public void saveEmailToken(String token) {
+		EmailToken emailToken = new EmailToken(token);
+		emailTokenRepository.save(emailToken);
+	}
 
 	// 인증된 토큰 삭제
 	@Transactional
 	public void deleteEmailToken(String token) {
-		emailTokenRepository.deleteById(token);
+		emailTokenRepository
+			.deleteById(token);
 	}
 
 	// 이메일 토큰 확인 및 삭제
@@ -29,6 +39,8 @@ public class EmailVerificationService {
 
 	// 토큰의 유효성을 확인하는 메서드
 	private boolean isTokenValid(String token) {
-		return emailTokenRepository.findById(token).isPresent();
+		return emailTokenRepository
+			.findById(token)
+			.isPresent();
 	}
 }
